refactor(suite-common): move `selectDeviceThunk` to `@suite-common/device`
