+++ packages/suite/src/components/suite/layouts/SuiteLayout/CoinjoinBars/CoinjoinStatusBar.tsx
@@ -4,9 +4,9 @@ import { selectRoundsDurationInHours, selectSessionProgressByAccountKey } from '
 import { type CoinjoinSession } from '@suite/coinjoin';
 import { Translation } from '@suite/intl';
 import { goto, selectRouterParams } from '@suite/router';
-import { selectDevices, selectSelectedDevice } from '@suite-common/device';
+import { selectDeviceThunk, selectDevices, selectSelectedDevice } from '@suite-common/device';
 import { useDispatch } from '@suite-common/redux-utils';
-import { selectAccountByKey, selectDeviceThunk } from '@suite-common/wallet-core';
+import { selectAccountByKey } from '@suite-common/wallet-core';
 import { type AccountKey, type WalletParams } from '@suite-common/wallet-types';
 import { ProgressPie } from '@trezor/components';
 import { typography } from '@trezor/theme';
