+++ packages/suite/src/components/suite/notifications/NotificationRenderer/TransactionRenderer.tsx
@@ -2,12 +2,11 @@ import { HiddenPlaceholder } from '@suite/discreet-mode';
 import { Translation } from '@suite/intl';
 import { openModal } from '@suite/modal';
 import { getTxAnchor, goto, selectRouteName, selectRouterApp } from '@suite/router';
-import { selectDevices, selectSelectedDevice } from '@suite-common/device';
+import { selectDeviceThunk, selectDevices, selectSelectedDevice } from '@suite-common/device';
 import { useDispatch } from '@suite-common/redux-utils';
 import {
     selectAccounts,
     selectBlockchainState,
-    selectDeviceThunk,
     selectTransactions,
 } from '@suite-common/wallet-core';
 import {
