+++ packages/suite/src/views/suite/SwitchDevice/NeedsAttentionBanner.tsx
@@ -1,12 +1,13 @@
 import { useDevice } from '@suite/device';
 import { Translation, type TranslationKey } from '@suite/intl';
 import { goto } from '@suite/router';
+import { selectDeviceThunk } from '@suite-common/device';
 import { useDispatch } from '@suite-common/redux-utils';
 import {
     type DeviceStatus as ConnectedDeviceStatus,
     type getStatus,
 } from '@suite-common/suite-utils';
-import { acquireDevice, selectDeviceThunk } from '@suite-common/wallet-core';
+import { acquireDevice } from '@suite-common/wallet-core';
 import { Banner, type BannerIntent } from '@trezor/components';
 import { exhaustive } from '@trezor/type-utils';
