+++ suite-common/device/src/index.ts
@@ -4,6 +4,7 @@ export type * from './deviceDeps';
 export * from './deviceReducer';
 export * from './deviceSelectors';
 export * from './deviceUtils';
+export * from './selectDeviceThunk';
 export * from './sortDevices';
 export * from './usePinHook';
 export { getIsIgnoredEntropyCheckError } from './services/getIsIgnoredEntropyCheckError';
