+++ suite-common/suite-sync/src/suiteSyncMiddleware.ts
@@ -1,10 +1,9 @@
 import { type UnknownAction, isAnyOf } from '@reduxjs/toolkit';
 
-import { deviceActions, isTrezorDeviceWithState } from '@suite-common/device';
+import { deviceActions, isTrezorDeviceWithState, selectDeviceThunk } from '@suite-common/device';
 import { type MessageSystemRootState } from '@suite-common/message-system';
 import { type WithServices, createMiddlewareWithExtraDeps } from '@suite-common/redux-utils';
 import { type SuiteSyncDep } from '@suite-common/suite-sync-types';
-import { selectDeviceThunk } from '@suite-common/wallet-core';
 
 import {
     type WithSuiteSyncAndDeviceState,
